update libft: update libft sources

+++ libft/ft_strchr.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 char	*ft_strchr(const char *s, int c)
 {
-	int				i;
+	size_t			i;
 
 	i = -1;
 	while (s[++i])
